requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.1.0
gunicorn==21.2.0